# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://localhost:8080/home
  description: Go Gin 框架示例项目 API 文档，包含用户认证、文章管理、标签管理等接口
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Demo API
  version: "1.0"
paths:
...
  /articles/create:
    get:
      description: 返回新增文章表单页面，包含标题、内容输入和标签选择功能
      responses:
        "200":
          description: 新增文章HTML
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 新增文章页面
      tags:
      - 文章管理
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 创建新文章，支持同时设置标签（逗号分隔或数组形式传入）
      parameters:
      - description: 文章标题
        in: formData
        name: title
        required: true
        type: string
      - description: 文章内容
        in: formData
        name: content
        required: true
        type: string
      - collectionFormat: csv
        description: 标签名称列表（可选）
        in: formData
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "302":
          description: 重定向到文章列表
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建文章
      tags:
      - 文章管理
  /articles/edit:
    get:
      description: 返回指定文章的编辑页面，预填充标题、内容和当前标签
      parameters:
      - description: 文章ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 编辑文章HTML
          schema:
            type: string
        "400":
          description: 无效的文章ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 编辑文章页面
      tags:
      - 文章管理
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 更新指定文章的标题、内容和标签。支持全量替换标签。
      parameters:
      - description: 文章ID
        in: formData
        name: id
        required: true
        type: integer
      - description: 文章标题
        in: formData
        name: title
        required: true
        type: string
      - description: 文章内容
        in: formData
        name: content
        required: true
        type: string
      - collectionFormat: csv
        description: 标签名称列表（全量替换）
        in: formData
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "302":
          description: 重定向到文章详情页
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新文章
      tags:
      - 文章管理
  /articles/list:
    get:
      description: 返回文章列表页面，支持按标签筛选。返回所有文章及其关联的作者和标签信息。
      parameters:
      - description: 按标签名称筛选
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文章列表HTML
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 文章列表
      tags:
      - 文章管理
  /articles/view:
    get:
      description: 返回指定文章的完整详情，包括标题、内容、作者和标签信息
      parameters:
      - description: 文章ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文章详情HTML
          schema:
            type: string
        "400":
          description: 无效的文章ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 文章详情
      tags:
      - 文章管理
...
  /home:
    get:
      description: 返回系统首页，展示用户信息和快捷入口
      responses:
        "200":
          description: 首页HTML
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 首页
      tags:
      - 文章管理
...
securityDefinitions:
  BearerAuth:
    description: '请在请求头中携带 Token: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"